refactor:#35 무한스크롤 로직 hook 형태로 변경
